Testing (#2)
* Updated documentation for ixyi

* Added KS cpp and documentation

* Added EBD_BR and EBD_RS_BR cpp and documentation

* Added srs cpp and documentation

* Added factorial_counts cpp and documentation
Added int2bin cpp and documentation
Added bin2int R and documentation

* Added factorial_time_bin_2 cpp and documentation.

* ixyi now starts the type indexing from 1.
Added factorial_time_bin for counting things for more than 2 alterantives. Returns matrix instead of list of matrices.

* for ixyi, changed ixyi_indicator to IntegerVector, but changed it back to std::vector<int> because too slow.

* Added ixyi_table_by_z for counting ixyi binned out by another variable. c++ and documentation.

* Had problems with factorial_time_bin, where the line:
IntegerVector time_bins = wrap( time/bin_resolution )
was causing crashes. We use arma::uvec instead.
Changed ixyi_table_by_z so that the bins are along the rows and categories along the columns. This is what it said in the documentation, but not in the actual code.

* Cleaned up documentation.
Fixed up arguments so they match the documentation and the code.
Fixed some @usage lines that were too long.